now with one actual check. time to test
